I believe I remember doing this to prepare for the interview, or during the interview itself.
The first step in this process is realizing that this needs to accep both arrays and objects. One of the key differences between the two are that you can iterate through an array using a for look while you need to use a for in look to iterate through an object.

I tried to use a ternary conditional, but it ended up getting to complicated and busting the whole thing. So i just used an if statement. As mentioned in the bootstrap i used Array.isArray to see whether or not the collection argument is an array, if it is, iterate through each item and perform the iterator on each index. Here the iterator function takes the value using bracket notation (collection[i]), the key (or index is i) and the collection is the same.

If the collection is an object, it fails the if statement and goes to the else statement, where we do the same iteration, but use the for in look to iterate through the object. Here I believe it is common convention to use key here.