_.filter was another function that I looked over and recreated extensively in preparation for the interview coding test.
This one seemed pretty easy, especially reusing the previously written _.each function. First, we can see that we need to pass in the collection (array or obj) and a callback to a truth test. Next we can initialize an empty array to collect the items that pass the test.

Next, we iterate over each of the items in the collection using _.each and passing a callback function, that when satisfied pushes a value onto the truths array. Finally we return this array.