I struggled with this for a while, implementing a few ways and commenting out one of the ways I couldn't get to work
Due to my time crunch this is a brute force way to put all the unique values into a new array. First initialized several variables that are used to store the unique items as well as store some temporary values and counters. We iterate through the given array and pass the value at each iteration into and object where we then test to see if a value is present using !== and if there is nothing there we insert a 1 as the value to the key, then we add that value to the array.