First, this function returns an array, thus we need to intialize an empty array to begin with
The basic components of this function are to apply a function to each item in an array and put the result of that index into a new array, creating a new array. Thus, we can use the _.each function that is already written and just plug the value at each index into whatever callback function is used as the iterator.