Getting tired, went around and deleted half worked on stuff because i coded in a rush and jumbled a lot of things. Refocused on one at a time here
First thing that needed to be done was to get all the arguments that were passed in after func and wait. I did a lot of googling and read up on call which allows us to call a function given a value and arguments individually. We used settimeout to execute the function after waiting a specific amount of time, this was similar to what I did in twittler.