Memoize was extremely confusing and it is getting very late
I've been having to use call/bind/apply pretty regularly afte looking at _.delay. Again here we need the arguments and we use slice and call to perform that and store that in args. then we check whether or not these arguments exist, it it does than we dont need to compute if it does not than we add it to the final object using apply to set the arguments.